Process per-project, split logic and plugin
