+++ delta/plugins/project-deletion/src/main/scala/ch/epfl/bluebrain/nexus/delta/plugins/projectdeletion/ProjectDeletionPlugin.scala
@@ -1,21 +1,19 @@
 package ch.epfl.bluebrain.nexus.delta.plugins.projectdeletion
 
-import cats.syntax.all._
-import ch.epfl.bluebrain.nexus.delta.kernel.utils.IOUtils
+import ch.epfl.bluebrain.nexus.delta.plugins.projectdeletion.ProjectDeletionLogic.projectDeletionPass
 import ch.epfl.bluebrain.nexus.delta.plugins.projectdeletion.model.ProjectDeletionConfig
 import ch.epfl.bluebrain.nexus.delta.sdk.ProjectResource
 import ch.epfl.bluebrain.nexus.delta.sdk.model.search.Pagination
 import ch.epfl.bluebrain.nexus.delta.sdk.model.search.SearchParams.ProjectSearchParams
 import ch.epfl.bluebrain.nexus.delta.sdk.plugin.Plugin
 import ch.epfl.bluebrain.nexus.delta.sdk.projects.{Projects, ProjectsStatistics}
-import ch.epfl.bluebrain.nexus.delta.sdk.syntax._
 import ch.epfl.bluebrain.nexus.delta.sourcing.model.Identity
 import ch.epfl.bluebrain.nexus.delta.sourcing.model.Identity.Subject
 import com.typesafe.scalalogging.Logger
 import fs2.Stream
-import monix.bio.{Fiber, IO, Task, UIO}
+import monix.bio.{Fiber, Task, UIO}
 
-import java.time.{Duration, Instant}
+import java.time.Instant
 import scala.concurrent.duration.DurationInt
 
 /**
@@ -39,79 +37,6 @@ class ProjectDeletionPlugin private (fiber: Fiber[Throwable, Unit]) extends Plug
 object ProjectDeletionPlugin {
 
   private val logger: Logger = Logger[ProjectDeletionPlugin]
-  implicit private class BooleanTaskOps(val left: Boolean) extends AnyVal {
-    def `<||>`(right: UIO[Boolean]): UIO[Boolean] = {
-      if (left) {
-        UIO.pure(true)
-      } else {
-        right
-      }
-    }
-
-    def `<&&>`(right: UIO[Boolean]): UIO[Boolean] = {
-      if (left) {
-        right
-      } else {
-        UIO.pure(false)
-      }
-    }
-  }
-
-  def projectDeletionPass(
-      allProjects: UIO[Seq[ProjectResource]],
-      deleteProject: ProjectResource => UIO[Unit],
-      config: ProjectDeletionConfig,
-      lastEventTime: (ProjectResource, Instant) => UIO[Instant]
-  ): UIO[Unit] = {
-
-    def isIncluded(pr: ProjectResource): Boolean = {
-      config.includedProjects.exists(regex => regex.matches(pr.value.ref.toString))
-    }
-
-    def notExcluded(pr: ProjectResource): Boolean = {
-      !config.excludedProjects.exists(regex => regex.matches(pr.value.ref.toString))
-    }
-
-    def deletableDueToDeprecation(pr: ProjectResource): Boolean = {
-      config.deleteDeprecatedProjects && pr.deprecated
-    }
-
-    def deletableDueToBeingIdle(pr: ProjectResource, now: Instant): UIO[Boolean] = {
-      projectIsIdle(pr, now) <&&> resourcesAreIdle(pr, now)
-    }
-
-    def projectIsIdle(pr: ProjectResource, now: Instant) = {
-      (now diff pr.updatedAt).toSeconds > config.idleInterval.toSeconds
-    }
-
-    def resourcesAreIdle(pr: ProjectResource, now: Instant): UIO[Boolean] = {
-      lastEventTime(pr, now).map(_.isBefore(now.minus(Duration.ofMillis(config.idleInterval.toMillis))))
-    }
-
-    def allProjectsNotAlreadyDeleted: IO[Nothing, Seq[ProjectResource]] = {
-      allProjects
-        .map(_.filter(!_.value.markedForDeletion))
-    }
-
-    def shouldBeDeleted(now: Instant): ProjectResource => UIO[Boolean] = { pr =>
-      {
-        (isIncluded(pr) && notExcluded(pr)) <&&> (deletableDueToDeprecation(pr) <||> deletableDueToBeingIdle(pr, now))
-      }
-    }
-
-    def deleteProjects(projects: Seq[ProjectResource]): UIO[Unit] = {
-      projects.traverse(deleteProject).void
-    }
-
-    for {
-      allProjects      <- allProjectsNotAlreadyDeleted
-      now              <- IOUtils.instant
-      projectsToDelete <- allProjects.filterA(shouldBeDeleted(now))
-      _                <- deleteProjects(projectsToDelete)
-    } yield {
-      ()
-    }
-  }
 
   /**
     * Constructs a stoppable ProjectDeletion process that is started in the background.
